restructure app to be easier to work on
