+++ logic/common/asyncForEach.js
@@ -1,0 +1,9 @@
+module.exports = (config, app, common) => {
+
+  return async function asyncForEach(array, callback) {
+    for (let index = 0; index < array.length; index++) {
+      await callback(array[index], index, array)
+    }
+  }
+
+}
