+++ logic/route/updateFile.js
@@ -1,0 +1,8 @@
+module.exports = (config, app, common, route) => {
+
+  // Update DB info for single file
+  return async function updateFile (req, res) {
+    return route.notImplemented(req, res)
+  }
+
+}
