+++ models/models.js
@@ -1,0 +1,7 @@
+module.exports = (app) => {
+  return {
+    auth: require('./auth.js')(app),
+    file: require('./file.js')(app),
+    album: require('./album.js')(app)
+  }
+}
